+++ lib/Dialect/Comb/CombFolds.cpp
@@ -186,7 +186,8 @@ getLowestBitAndHighestBitRequired(Operation *op, bool narrowTrailingBits,
   for (auto *user : users) {
     if (auto extractOp = dyn_cast<ExtractOp>(user)) {
       size_t lowBit = extractOp.getLowBit();
-      size_t highBit = extractOp.getType().getWidth() + lowBit - 1;
+      size_t highBit =
+          extractOp.getType().cast<IntegerType>().getWidth() + lowBit - 1;
       highestBitRequired = std::max(highestBitRequired, highBit);
       lowestBitRequired = std::min(lowestBitRequired, lowBit);
       continue;
@@ -245,15 +246,20 @@ static bool narrowOperationWidth(OpTy op, bool narrowTrailingBits,
 
 OpFoldResult ReplicateOp::fold(ArrayRef<Attribute> constants) {
   // Replicate one time -> noop.
-  if (getType().getWidth() == getInput().getType().getIntOrFloatBitWidth())
+  if (getType().cast<IntegerType>().getWidth() ==
+      getInput().getType().getIntOrFloatBitWidth())
     return getInput();
 
   // Constant fold.
   if (auto input = constants[0].dyn_cast_or_null<IntegerAttr>()) {
     if (input.getValue().getBitWidth() == 1) {
       if (input.getValue().isZero())
-        return getIntAttr(APInt::getZero(getType().getWidth()), getContext());
-      return getIntAttr(APInt::getAllOnes(getType().getWidth()), getContext());
+        return getIntAttr(
+            APInt::getZero(getType().cast<IntegerType>().getWidth()),
+            getContext());
+      return getIntAttr(
+          APInt::getAllOnes(getType().cast<IntegerType>().getWidth()),
+          getContext());
     }
 
     APInt result = APInt::getZeroWidth();
@@ -415,7 +421,7 @@ OpFoldResult ExtractOp::fold(ArrayRef<Attribute> constants) {
 
   // Constant fold.
   if (auto input = constants[0].dyn_cast_or_null<IntegerAttr>()) {
-    unsigned dstWidth = getType().getWidth();
+    unsigned dstWidth = getType().cast<IntegerType>().getWidth();
     return getIntAttr(input.getValue().lshr(getLowBit()).trunc(dstWidth),
                       getContext());
   }
@@ -471,7 +477,7 @@ static LogicalResult extractConcatToConcatExtract(ExtractOp op,
          "lowBit");
 
   SmallVector<Value> reverseConcatArgs;
-  size_t widthRemaining = op.getType().getWidth();
+  size_t widthRemaining = op.getType().cast<IntegerType>().getWidth();
   size_t extractLo = lowBit - beginOfFirstRelevantElement;
 
   // Transform individual arguments of innerCat(..., a, b, c,) into
@@ -508,7 +514,7 @@ static LogicalResult extractConcatToConcatExtract(ExtractOp op,
 // Transforms extract(lo, replicate(a, N)) into replicate(a, N-c).
 static bool extractFromReplicate(ExtractOp op, ReplicateOp replicate,
                                  PatternRewriter &rewriter) {
-  auto extractResultWidth = op.getType().getWidth();
+  auto extractResultWidth = op.getType().cast<IntegerType>().getWidth();
   auto replicateEltWidth =
       replicate.getOperand().getType().getIntOrFloatBitWidth();
 
@@ -541,7 +547,8 @@ LogicalResult ExtractOp::canonicalize(ExtractOp op, PatternRewriter &rewriter) {
 
   // If the extracted bits are all known, then return the result.
   auto knownBits = computeKnownBits(op.getInput())
-                       .extractBits(op.getType().getWidth(), op.getLowBit());
+                       .extractBits(op.getType().cast<IntegerType>().getWidth(),
+                                    op.getLowBit());
   if (knownBits.isConstant()) {
     replaceOpWithNewOpAndCopyName<hw::ConstantOp>(rewriter, op,
                                                   knownBits.getConstant());
@@ -570,8 +577,8 @@ LogicalResult ExtractOp::canonicalize(ExtractOp op, PatternRewriter &rewriter) {
   if (inputOp && inputOp->getNumOperands() == 2 &&
       isa<AndOp, OrOp, XorOp>(inputOp)) {
     if (auto cstRHS = inputOp->getOperand(1).getDefiningOp<hw::ConstantOp>()) {
-      auto extractedCst = cstRHS.getValue().extractBits(op.getType().getWidth(),
-                                                        op.getLowBit());
+      auto extractedCst = cstRHS.getValue().extractBits(
+          op.getType().cast<IntegerType>().getWidth(), op.getLowBit());
       if (isa<OrOp, XorOp>(inputOp) && extractedCst.isZero()) {
         replaceOpWithNewOpAndCopyName<ExtractOp>(
             rewriter, op, op.getType(), inputOp->getOperand(0), op.getLowBit());
@@ -609,7 +616,7 @@ LogicalResult ExtractOp::canonicalize(ExtractOp op, PatternRewriter &rewriter) {
 
   // `extract(lowBit, shl(1, x))` -> `x == lowBit` when a single bit is
   // extracted.
-  if (op.getType().getWidth() == 1 && inputOp)
+  if (op.getType().cast<IntegerType>().getWidth() == 1 && inputOp)
     if (auto shlOp = dyn_cast<ShlOp>(inputOp))
       if (auto lhsCst = shlOp.getOperand(0).getDefiningOp<hw::ConstantOp>())
         if (lhsCst.getValue().isOne()) {
@@ -742,7 +749,7 @@ static bool canonicalizeLogicalCstWithConcat(Operation *logicalOp,
 }
 
 OpFoldResult AndOp::fold(ArrayRef<Attribute> constants) {
-  APInt value = APInt::getAllOnes(getType().getWidth());
+  APInt value = APInt::getAllOnes(getType().cast<IntegerType>().getWidth());
 
   // and(x, 01, 10) -> 00 -- annulment.
   for (auto operand : constants) {
@@ -769,7 +776,9 @@ OpFoldResult AndOp::fold(ArrayRef<Attribute> constants) {
     if (matchPattern(arg, m_Complement(m_Any(&subExpr)))) {
       for (Value arg2 : getInputs())
         if (arg2 == subExpr)
-          return getIntAttr(APInt::getZero(getType().getWidth()), getContext());
+          return getIntAttr(
+              APInt::getZero(getType().cast<IntegerType>().getWidth()),
+              getContext());
     }
   }
 
@@ -923,7 +932,7 @@ LogicalResult AndOp::canonicalize(AndOp op, PatternRewriter &rewriter) {
 }
 
 OpFoldResult OrOp::fold(ArrayRef<Attribute> constants) {
-  auto value = APInt::getZero(getType().getWidth());
+  auto value = APInt::getZero(getType().cast<IntegerType>().getWidth());
 
   // or(x, 10, 01) -> 11
   for (auto operand : constants) {
@@ -950,8 +959,9 @@ OpFoldResult OrOp::fold(ArrayRef<Attribute> constants) {
     if (matchPattern(arg, m_Complement(m_Any(&subExpr)))) {
       for (Value arg2 : getInputs())
         if (arg2 == subExpr)
-          return getIntAttr(APInt::getAllOnes(getType().getWidth()),
-                            getContext());
+          return getIntAttr(
+              APInt::getAllOnes(getType().cast<IntegerType>().getWidth()),
+              getContext());
     }
   }
 
@@ -1393,7 +1403,7 @@ OpFoldResult MulOp::fold(ArrayRef<Attribute> constants) {
   if (size == 1u)
     return getInputs()[0];
 
-  auto width = getType().getWidth();
+  auto width = getType().cast<IntegerType>().getWidth();
   APInt value(/*numBits=*/width, 1, /*isSigned=*/false);
 
   // mul(x, 0, 1) -> 0 -- annulment
@@ -1630,7 +1640,7 @@ LogicalResult ConcatOp::canonicalize(ConcatOp op, PatternRewriter &rewriter) {
       if (auto extract = inputs[i].getDefiningOp<ExtractOp>()) {
         if (auto prevExtract = inputs[i - 1].getDefiningOp<ExtractOp>()) {
           if (extract.getInput() == prevExtract.getInput()) {
-            auto thisWidth = extract.getType().getWidth();
+            auto thisWidth = extract.getType().cast<IntegerType>().getWidth();
             if (prevExtract.getLowBit() == extract.getLowBit() + thisWidth) {
               auto prevWidth = prevExtract.getType().getIntOrFloatBitWidth();
               auto resType = rewriter.getIntegerType(thisWidth + prevWidth);
